Reorganize E2E tests and add a project reference test
