+++ eng/pipelines/pr.yml
@@ -55,7 +55,7 @@ jobs:
     displayName: E2E tests - Pack NuGet.Tests library
     inputs:
       command: pack
-      packagesToPack: tests/NuGet.Tests/NuGetWithBasePipeline/NuGetWithBasePipeline.csproj
+      packagesToPack: tests/NuGet.Tests/SharplinerLibrary/SharplinerLibrary.csproj
       configurationToPack: Release
       outputDir: artifacts/packages
       workingDirectory: tests/NuGet.Tests
