+++ tests/E2E.Tests/NuGet.config
