+++ tests/E2E.Tests/SharplinerLibrary/BasePipelineFromLibrary.cs
@@ -1,7 +1,6 @@
 using Sharpliner.AzureDevOps;
-using Sharpliner.AzureDevOps.Tasks;
 
-namespace Tests.NuGetWithBasePipeline;
+namespace E2E.Tests.SharplinerLibrary;
 
 // These NuGet.Tests are E2E testing following scenario:
 // 1. We create a "library" project with an arbitrary Sharpliner definition
@@ -12,7 +11,7 @@ namespace Tests.NuGetWithBasePipeline;
 // This is testing user scenarios where people want to publish definitions in the form of NuGet and re-use it in other projects.
 
 // This class is the one that goes into the NuGet
-public abstract class BasePipelineFromNuGet : SingleStagePipelineDefinition
+public abstract class BasePipelineFromLibrary : SingleStagePipelineDefinition
 {
     public override SingleStagePipeline Pipeline => new()
     {
