+++ tests/E2E.Tests/SharplinerLibrary/E2E.Tests.SharplinerLibrary.csproj
